% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{nc.get.dim.names}
\alias{nc.get.dim.names}
\title{Get a list of names of dimensions}
\usage{
nc.get.dim.names(f, v)
}
\arguments{
\item{f}{The file (an object of class \code{ncdf4})}

\item{v}{Optionally, a variable}
}
\value{
A character vector naming the dimensions found.
}
\description{
Get a list of names of dimensions.
}
\details{
This function returns the names of dimensions in a file or, if \code{v} is also supplied, attached to a particular variable.
}
\examples{
## Get dimension names
\dontrun{
f <- nc_open("pr.nc")
dim.names <- nc.get.dim.names(f, "pr")
nc_close(f)
}

}
