% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{nc.get.time.multiplier}
\alias{nc.get.time.multiplier}
\title{Gets conversion factor for time scale given units}
\usage{
nc.get.time.multiplier(x)
}
\arguments{
\item{x}{The time scale}
}
\value{
A numeric conversion factor to convert to seconds.
}
\description{
Gets conversion factor for time scale given units.
}
\details{
This function returns a conversion factor from the supplied time scale (days, hours, minutes, months) to seconds. This can be used to convert to/from "(days or hours) since X" style dates.
}
\note{
The conversion factor for months is approximate.
}
\examples{
## Will return 3600
mul <- nc.get.time.multiplier("hours")

}
